Add HeaderNotificationMenu
