+++ src/components/Header.js
@@ -1,24 +1,14 @@
 import Link from 'next/link'
-import { useState, useRef } from 'react'
-import request from '../utils/request'
-import useSWR from 'swr'
 import { makeStyles } from '@material-ui/core/styles'
 import AppBar from '@material-ui/core/AppBar'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
 import Button from '@material-ui/core/Button'
-import IconButton from '@material-ui/core/IconButton'
-import Badge from '@material-ui/core/Badge'
-import Menu from '@material-ui/core/Menu'
-import MenuItem from '@material-ui/core/MenuItem'
-import NotificationsIcon from '@material-ui/icons/Notifications'
 import firebase from '../utils/Firebase'
+import HeaderNotificationMenu from './HeaderNotificationMenu'
 
 export default function Header({ user, loading }) {
   const classes = useStyles()
-  const [isNotificationOpen, setIsNotificationOpen] = useState(false)
-  const notificationAnchorRef = useRef(null)
-  const { data: notifications } = useSWR('/me/notifications', notificationsFetcher)
 
   const logOut = async () => {
     try {
@@ -26,14 +16,6 @@ export default function Header({ user, loading }) {
     } catch (error) {
       console.log(error)
     }
-  }
-
-  const handleNotificationOpen = () => {
-    setIsNotificationOpen(true)
-  }
-
-  const handleNotificationClose = () => {
-    setIsNotificationOpen(false)
   }
 
   return (
@@ -50,36 +32,7 @@ export default function Header({ user, loading }) {
               <Link href='/projects/new'>
                 <Button color='inherit'>はじめる</Button>
               </Link>
-              <IconButton color='inherit' onClick={handleNotificationOpen} ref={notificationAnchorRef}>
-                <Badge badgeContent={notifications && notifications.length} color='secondary'>
-                  <NotificationsIcon />
-                </Badge>
-              </IconButton>
-              <Menu
-                anchorEl={notificationAnchorRef.current}
-                keepMounted
-                getContentAnchorEl={null}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'left',
-                }}
-                open={isNotificationOpen}
-                onClose={handleNotificationClose}
-              >
-                {notifications && notifications.length > 0 ? (
-                  notifications.map(notification => (
-                    <MenuItem key={notification.id}>
-                      {notification.link !== null ? (
-                        <Link href={notification.link}>{notification.body}</Link>
-                      ) : (
-                        notification.body
-                      )}
-                    </MenuItem>
-                  ))
-                ) : (
-                  <MenuItem>通知はありません</MenuItem>
-                )}
-              </Menu>
+              <HeaderNotificationMenu />
               <Link href='/profile'>
                 <Button color='inherit'>{user.name}</Button>
               </Link>
@@ -101,8 +54,6 @@ export default function Header({ user, loading }) {
     </AppBar>
   )
 }
-
-const notificationsFetcher = url => request.get(url).then(res => res.data.notifications)
 
 const useStyles = makeStyles(theme => ({
   root: {
